\name{keboola.r.luckyguess-package}
\alias{keboola.r.luckyguess-package}
\alias{keboola.r.luckyguess}
\docType{package}
\title{
R backend for LuckyGuess component in Keboola Connection
}
\description{
Package is an R application which executes an R script supplied in configuration parameters.
The configuration format is defined by docker-bundle, the R script is expected to directly 
communicate with Amazon Redshift backend.
}
\details{
\tabular{ll}{
Package: \tab keboola.r.luckyguess\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-04\cr
License: \tab MIT\cr
}
}
\author{
Maintainer: Ondrej Popelka <ondrej.popelka@keboola.com>
}\keyword{ programming, database }
\examples{
# TODO
}
