% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keboola.r.luckyguess.R
\docType{class}
\name{LGApplication-class}
\alias{LGApplication}
\alias{LGApplication-class}
\title{Application which serves as a backend for LuckyGuess component}
\description{
Application which serves as a backend for LuckyGuess component
}

